# Generalize optimizer step profiler