--- a/libyul/optimiser/Suite.cpp
+++ b/libyul/optimiser/Suite.cpp
@@ -72,6 +72,7 @@
 #include <libyul/backends/evm/NoOutputAssembly.h>
 
 #include <libsolutil/CommonData.h>
+#include <libsolutil/Profiler.h>
 
 #include <libyul/CompilabilityChecker.h>
 
@@ -83,55 +84,9 @@
 #include <limits>
 #include <tuple>
 
-#ifdef PROFILE_OPTIMIZER_STEPS
-#include <chrono>
-#include <fmt/format.h>
-#endif
-
 using namespace solidity;
 using namespace solidity::yul;
-#ifdef PROFILE_OPTIMIZER_STEPS
-using namespace std::chrono;
-#endif
 using namespace std::string_literals;
-
-namespace
-{
-
-#ifdef PROFILE_OPTIMIZER_STEPS
-void outputPerformanceMetrics(std::map<std::string, int64_t> const& _metrics)
-{
-	std::vector<std::pair<std::string, int64_t>> durations(_metrics.begin(), _metrics.end());
-	sort(
-		durations.begin(),
-		durations.end(),
-		[](std::pair<std::string, int64_t> const& _lhs, std::pair<std::string, int64_t> const& _rhs) -> bool
-		{
-			return _lhs.second < _rhs.second;
-		}
-	);
-
-	int64_t totalDurationInMicroseconds = 0;
-	for (auto&& [step, durationInMicroseconds]: durations)
-		totalDurationInMicroseconds += durationInMicroseconds;
-
-	std::cerr << "Performance metrics of optimizer steps" << std::endl;
-	std::cerr << "======================================" << std::endl;
-	constexpr double microsecondsInSecond = 1000000;
-	for (auto&& [step, durationInMicroseconds]: durations)
-	{
-		double percentage = 100.0 * static_cast<double>(durationInMicroseconds) / static_cast<double>(totalDurationInMicroseconds);
-		double sec = static_cast<double>(durationInMicroseconds) / microsecondsInSecond;
-		std::cerr << fmt::format("{:>7.3f}% ({} s): {}", percentage, sec, step) << std::endl;
-	}
-	double totalDurationInSeconds = static_cast<double>(totalDurationInMicroseconds) / microsecondsInSecond;
-	std::cerr << "--------------------------------------" << std::endl;
-	std::cerr << fmt::format("{:>7}% ({:.3f} s)", 100, totalDurationInSeconds) << std::endl;
-}
-#endif
-
-}
-
 
 void OptimiserSuite::run(
 	Dialect const& _dialect,
@@ -225,10 +180,6 @@ void OptimiserSuite::run(
 	dispenser.reset(astRoot);
 	NameSimplifier::run(suite.m_context, astRoot);
 	VarNameCleaner::run(suite.m_context, astRoot);
-
-#ifdef PROFILE_OPTIMIZER_STEPS
-	outputPerformanceMetrics(suite.m_durationPerStepInMicroseconds);
-#endif
 
 	_object.setCode(std::make_shared<AST>(std::move(astRoot)));
 	_object.analysisInfo = std::make_shared<AsmAnalysisInfo>(AsmAnalyzer::analyzeStrictAssertCorrect(_dialect, _object));
@@ -504,14 +455,12 @@ void OptimiserSuite::runSequence(std::vector<std::string> const& _steps, Block& 
 	{
 		if (m_debug == Debug::PrintStep)
 			std::cout << "Running " << step << std::endl;
-#ifdef PROFILE_OPTIMIZER_STEPS
-		steady_clock::time_point startTime = steady_clock::now();
-#endif
-		allSteps().at(step)->run(m_context, _ast);
-#ifdef PROFILE_OPTIMIZER_STEPS
-		steady_clock::time_point endTime = steady_clock::now();
-		m_durationPerStepInMicroseconds[step] += duration_cast<microseconds>(endTime - startTime).count();
-#endif
+
+		{
+			PROFILER_PROBE(step, probe);
+			allSteps().at(step)->run(m_context, _ast);
+		}
+
 		if (m_debug == Debug::PrintChanges)
 		{
 			// TODO should add switch to also compare variable names!