--- a/libyul/optimiser/Suite.h
+++ b/libyul/optimiser/Suite.h
@@ -91,9 +91,6 @@ public:
 private:
 	OptimiserStepContext& m_context;
 	Debug m_debug;
-#ifdef PROFILE_OPTIMIZER_STEPS
-	std::map<std::string, int64_t> m_durationPerStepInMicroseconds;
-#endif
 };
 
 }