This PR moves the simple profiling setup we had in `Suite.cpp` and moves it into a separate class that can be easily used in random places.

It also switches to static storage for the metrics, which means that we get a single report with total running time over all the Yul objects rather than one per object. Originally the profiler was used with small contracts from `test/benchmarks/` where it was probably manageable, but when profiling big projects like Uniswap or OpenZeppelin, the aggregate data is much more useful.

Using the profiler is now as easy as dropping a `PROFILER_PROBE()` macro in the scope that is meant to be measured. The default probes are still placed only around the optimizer steps, but new ones can be added ad-hoc, locally.

The PR also tweaks the output to make it easy to post it in comments in a readable form. It also tracks the number of calls in addition to the timing.