@cameel 2024-10-11
So, as can be seen in [run 36438](https://app.circleci.com/pipelines/github/ethereum/solidity/36438/workflows/435472c3-9937-4f28-9316-b448ed0d3874), the code does compile on all platforms.

The three CLI test jobs fail, but that was already the case before due to the extra output that the profiler prints unconditionally. Perhaps we should add a proper flag for it, but in this PR I'm just leaving it as is.